+++ src/main/java/org/folio/persist/AbstractRepository.java
@@ -82,7 +82,19 @@ public abstract class AbstractRepository<T> {
     return promise.future();
   }
 
+  public Future<RowSet<Row>> update(List<T> records) {
+    final Promise<RowSet<Row>> promise = promise();
+
+    postgresClient.upsertBatch(tableName, records, promise);
+
+    return promise.future();
+  }
+
   public Future<RowSet<Row>> deleteAll() {
     return postgresClientFuturized.delete(tableName, new Criterion());
+  }
+
+  public Future<RowSet<Row>> deleteById(String id) {
+    return postgresClientFuturized.deleteById(tableName, id);
   }
 }