+++ src/main/java/org/folio/rest/persist/PostgresClientFuturized.java
@@ -34,28 +34,12 @@ public class PostgresClientFuturized {
     return saveResult.future();
   }
 
-  public <T> Future<T> saveAndReturnEntity(String table, String id, T entity) {
-    final Promise<T> saveResult = promise();
-
-    postgresClient.saveAndReturnUpdatedEntity(table, id, entity, saveResult);
-
-    return saveResult.future();
-  }
-
   public <T> Future<T> getById(String tableName, String id, Class<T> type) {
     final Promise<T> getByIdResult = promise();
 
     postgresClient.getById(tableName, id, type, getByIdResult);
 
     return getByIdResult.future();
-  }
-
-  public <T> Future<List<T>> get(String tableName, T object) {
-    final Promise<Results<T>> getAllItemsResult = promise();
-
-    postgresClient.get(tableName, object, false, getAllItemsResult);
-
-    return getAllItemsResult.future().map(Results::getResults);
   }
 
   public <T> Future<List<T>> get(String tableName, Class<T> type, Criterion criterion) {
@@ -72,6 +56,14 @@ public class PostgresClientFuturized {
     postgresClient.delete(tableName, criterion, removeAllResult);
 
     return removeAllResult.future();
+  }
+
+  public Future<RowSet<Row>> deleteById(String tableName, String id) {
+    final Promise<RowSet<Row>> deleteResult = promise();
+
+    postgresClient.delete(tableName, id, deleteResult);
+
+    return deleteResult.future();
   }
 
   public <T> Future<Map<String, T>> getById(String tableName, Collection<String> ids, Class<T> type) {