+++ src/main/java/org/folio/services/domainevent/CommonDomainEventPublisher.java
@@ -3,53 +3,59 @@ package org.folio.services.domainevent;
 import static io.vertx.core.CompositeFuture.all;
 import static io.vertx.core.Future.succeededFuture;
 import static org.apache.logging.log4j.LogManager.getLogger;
-import static org.folio.okapi.common.XOkapiHeaders.TENANT;
-import static org.folio.okapi.common.XOkapiHeaders.URL;
 import static org.folio.rest.tools.utils.TenantTool.tenantId;
 import static org.folio.services.domainevent.DomainEvent.createEvent;
 import static org.folio.services.domainevent.DomainEvent.deleteAllEvent;
 import static org.folio.services.domainevent.DomainEvent.deleteEvent;
 import static org.folio.services.domainevent.DomainEvent.updateEvent;
-import static org.folio.services.kafka.KafkaProducerServiceFactory.getKafkaProducerService;
 
 import io.vertx.core.Context;
 import io.vertx.core.Future;
+import io.vertx.core.Promise;
+import io.vertx.core.streams.ReadStream;
+import io.vertx.kafka.client.producer.KafkaProducer;
 import java.util.Collection;
-import java.util.Collections;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
+import java.util.concurrent.atomic.AtomicLong;
+import java.util.function.Function;
 import java.util.stream.Collectors;
 import org.apache.commons.lang3.tuple.Pair;
 import org.apache.commons.lang3.tuple.Triple;
 import org.apache.logging.log4j.Logger;
-import org.folio.services.kafka.KafkaMessage;
+import org.folio.kafka.KafkaConfig;
+import org.folio.kafka.KafkaProducerManager;
+import org.folio.kafka.SimpleKafkaProducerManager;
+import org.folio.services.kafka.InventoryProducerRecordBuilder;
+import org.folio.services.kafka.KafkaProperties;
 import org.folio.services.kafka.topic.KafkaTopic;
 
 public class CommonDomainEventPublisher<T> {
-  private static final Set<String> FORWARDER_HEADERS = Set.of(URL.toLowerCase(),
-    TENANT.toLowerCase());
   public static final String NULL_INSTANCE_ID = "00000000-0000-0000-0000-000000000000";
   private static final Logger log = getLogger(CommonDomainEventPublisher.class);
 
-  private final Context vertxContext;
-  private final Map<String, String> headersToForward;
+  private final Map<String, String> okapiHeaders;
   private final KafkaTopic kafkaTopic;
-  protected final String tenantId;
+  private final KafkaProducerManager producerManager;
 
-  CommonDomainEventPublisher(Context vertxContext, Map<String, String> okapiHeaders,
+  CommonDomainEventPublisher(Map<String, String> okapiHeaders, KafkaTopic kafkaTopic,
+    KafkaProducerManager kafkaProducerManager) {
+
+    this.okapiHeaders = okapiHeaders;
+    this.kafkaTopic = kafkaTopic;
+    this.producerManager = kafkaProducerManager;
+  }
+
+  public CommonDomainEventPublisher(Context vertxContext, Map<String, String> okapiHeaders,
     KafkaTopic kafkaTopic) {
 
-    this.vertxContext = vertxContext;
-    this.headersToForward = getHeadersToForward(okapiHeaders);
-    this.kafkaTopic = kafkaTopic;
-    this.tenantId = tenantId(headersToForward);
+    this(okapiHeaders, kafkaTopic, createProducerManager(vertxContext));
   }
 
   Future<Void> publishRecordUpdated(String instanceId, T oldRecord, T newRecord) {
-    final DomainEvent<T> domainEvent = updateEvent(oldRecord, newRecord, tenantId);
+    final DomainEvent<T> domainEvent = updateEvent(oldRecord, newRecord, tenantId(okapiHeaders));
 
-    return publishMessage(instanceId, domainEvent);
+    return publish(instanceId, domainEvent);
   }
 
   Future<Void> publishRecordsUpdated(Collection<Triple<String, T, T>> updatedRecords) {
@@ -64,9 +70,9 @@ public class CommonDomainEventPublisher<T> {
   }
 
   Future<Void> publishRecordCreated(String instanceId, T newRecord) {
-    final DomainEvent<T> domainEvent = createEvent(newRecord, tenantId);
+    final DomainEvent<T> domainEvent = createEvent(newRecord, tenantId(okapiHeaders));
 
-    return publishMessage(instanceId, domainEvent);
+    return publish(instanceId, domainEvent);
   }
 
   Future<Void> publishRecordsCreated(List<Pair<String, T>> records) {
@@ -81,44 +87,91 @@ public class CommonDomainEventPublisher<T> {
   }
 
   Future<Void> publishRecordRemoved(String instanceId, T oldEntity) {
-    final DomainEvent<T> domainEvent = deleteEvent(oldEntity, tenantId);
+    final DomainEvent<T> domainEvent = deleteEvent(oldEntity, tenantId(okapiHeaders));
 
-    return publishMessage(instanceId, domainEvent);
+    return publish(instanceId, domainEvent);
   }
 
   Future<Void> publishAllRecordsRemoved() {
-    return publishMessage(NULL_INSTANCE_ID, deleteAllEvent(tenantId));
+    return publish(NULL_INSTANCE_ID, deleteAllEvent(tenantId(okapiHeaders)));
   }
 
-  private Future<Void> publishMessage(String instanceId, DomainEvent<T> domainEvent) {
-    return publishMessage(instanceId, domainEvent, Collections.emptyMap());
+  public <R> Future<Long> publishStream(ReadStream<R> readStream,
+    Function<R, InventoryProducerRecordBuilder> mapper,
+    Function<Long, Future<?>> progressHandler) {
+
+    var promise = Promise.<Long>promise();
+    var kafkaProducer = getOrCreateProducer("stream_");
+    var recordsProcessed = new AtomicLong(0);
+
+    readStream.exceptionHandler(error -> {
+      log.error("Unable to publish stream", error);
+      promise.tryFail(error);
+    }).endHandler(notUsed -> promise.tryComplete(recordsProcessed.get()))
+      .handler(record -> {
+        var producerRecord = mapper.apply(record)
+          .topic(kafkaTopic).propagateOkapiHeaders(okapiHeaders).build();
+
+        kafkaProducer.send(producerRecord)
+          .onFailure(error -> {
+            log.error("Unable to send event [{}]", producerRecord.value(), error);
+            recordsProcessed.decrementAndGet();
+          });
+
+        progressHandler.apply(recordsProcessed.incrementAndGet())
+          .onFailure(error -> {
+            log.warn("Error occurred when progress tracked", error);
+
+            readStream.pause();
+            promise.tryFail(error);
+          });
+
+        if (kafkaProducer.writeQueueFull()) {
+          log.info("Producer write queue full...");
+          readStream.pause();
+
+          kafkaProducer.drainHandler(notUsed -> {
+            log.info("Producer write queue empty again...");
+            readStream.resume();
+          });
+        }
+      });
+
+    return promise.future()
+      .onSuccess(records -> log.info("Total records published from stream {}", records));
   }
 
-  Future<Void> publishMessage(String instanceId, DomainEvent<T> domainEvent,
-    Map<String, String> additionalHeaders) {
+  private Future<Void> publish(String key, DomainEvent<T> value) {
+    log.debug("Sending domain event [{}], payload [{}]", key, value);
 
-    log.debug("Sending domain event [{}], payload [{}]", instanceId, domainEvent);
+    var producerRecord = new InventoryProducerRecordBuilder()
+      .key(key).value(value).topic(kafkaTopic).propagateOkapiHeaders(okapiHeaders)
+      .build();
 
-    var kafkaMessage = KafkaMessage.builder()
-      .key(instanceId).payload(domainEvent)
-      .topic(kafkaTopic)
-      .headers(headersToForward);
-
-    additionalHeaders.forEach(kafkaMessage::header);
-
-    return getKafkaProducerService(vertxContext.owner())
-      .sendMessage(kafkaMessage.build())
+    return getOrCreateProducer().send(producerRecord)
+      .<Void>map(notUsed -> null)
       .onComplete(result -> {
         if (result.failed()) {
           log.error("Unable to send domain event [{}], payload - [{}]",
-            instanceId, domainEvent, result.cause());
+            key, value, result.cause());
         }
       });
   }
 
-  private static Map<String, String> getHeadersToForward(Map<String, String> okapiHeaders) {
-    return okapiHeaders.entrySet().stream()
-      .filter(entry -> FORWARDER_HEADERS.contains(entry.getKey().toLowerCase()))
-      .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
+  private KafkaProducer<String, String> getOrCreateProducer() {
+    return getOrCreateProducer("");
+  }
+
+  private KafkaProducer<String, String> getOrCreateProducer(String prefix) {
+    return producerManager.createShared(prefix + kafkaTopic.getTopicName());
+  }
+
+  private static KafkaProducerManager createProducerManager(Context vertxContext) {
+    var kafkaConfig = KafkaConfig.builder()
+      .kafkaPort(KafkaProperties.getPort())
+      .kafkaHost(KafkaProperties.getHost())
+      .build();
+
+    return new SimpleKafkaProducerManager(vertxContext.owner(), kafkaConfig);
   }
 }