+++ src/main/java/org/folio/services/domainevent/InstanceDomainEventPublisher.java
@@ -2,52 +2,28 @@ package org.folio.services.domainevent;
 
 import static io.vertx.core.Future.succeededFuture;
 import static org.apache.logging.log4j.LogManager.getLogger;
-import static org.folio.rest.support.ResponseUtil.isCreateSuccessResponse;
-import static org.folio.services.domainevent.DomainEvent.reindexEvent;
 import static org.folio.services.kafka.topic.KafkaTopic.INVENTORY_INSTANCE;
-
-import java.util.Collection;
-import java.util.List;
-import java.util.Map;
-import java.util.function.Function;
-import java.util.stream.Collectors;
-
-import javax.ws.rs.core.Response;
-
-import org.apache.commons.lang3.tuple.ImmutablePair;
-import org.apache.commons.lang3.tuple.ImmutableTriple;
-import org.apache.commons.lang3.tuple.Pair;
-import org.apache.commons.lang3.tuple.Triple;
-import org.apache.logging.log4j.Logger;
-import org.folio.persist.InstanceRepository;
-import org.folio.rest.jaxrs.model.Instance;
-import org.folio.services.batch.BatchOperationContext;
 
 import io.vertx.core.Context;
 import io.vertx.core.Future;
+import java.util.Collection;
+import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
+import org.apache.commons.lang3.tuple.Pair;
+import org.apache.logging.log4j.Logger;
+import org.folio.persist.InstanceRepository;
+import org.folio.rest.jaxrs.model.Instance;
 
-public class InstanceDomainEventPublisher {
+public class InstanceDomainEventPublisher extends AbstractDomainEventPublisher<Instance, Instance> {
   private static final Logger log = getLogger(InstanceDomainEventPublisher.class);
-  public static final String REINDEX_JOB_ID_HEADER = "reindex-job-id";
-
-  private final InstanceRepository instanceRepository;
-  private final CommonDomainEventPublisher<Instance> domainEventService;
 
   public InstanceDomainEventPublisher(Context context, Map<String, String> okapiHeaders) {
-    instanceRepository = new InstanceRepository(context, okapiHeaders);
-    domainEventService = new CommonDomainEventPublisher<>(context, okapiHeaders,
-      INVENTORY_INSTANCE);
+    super(new InstanceRepository(context, okapiHeaders),
+      new CommonDomainEventPublisher<>(context, okapiHeaders, INVENTORY_INSTANCE));
   }
 
-  public Future<Void> publishInstanceUpdated(Instance oldRecord, Instance newRecord) {
-    return domainEventService.publishRecordUpdated(oldRecord.getId(), oldRecord, newRecord);
-  }
-
-  public Future<Void> publishInstanceCreated(Instance newInstance) {
-    return domainEventService.publishRecordCreated(newInstance.getId(), newInstance);
-  }
-
-  public Future<Void> publishInstancesCreated(Collection<Instance> instances) {
+  public Future<Void> publishInstancesCreated(List<Instance> instances) {
     if (instances.isEmpty()) {
       log.info("No instances were created, skipping event sending");
       return succeededFuture();
@@ -55,64 +31,25 @@ public class InstanceDomainEventPublisher {
 
     log.info("[{}] instances were created, sending events for them", instances.size());
 
-    final var createdInstancesPairs = instances.stream()
-      .map(instance -> new ImmutablePair<>(instance.getId(), instance))
-      .collect(Collectors.<Pair<String, Instance>>toList());
-
-    return domainEventService.publishRecordsCreated(createdInstancesPairs);
+    return domainEventService.publishRecordsCreated(instances.stream()
+      .map(instance -> pair(instance.getId(), instance))
+      .collect(Collectors.toList()));
   }
 
-  public Future<Void> publishInstanceRemoved(Instance oldEntity) {
-    return domainEventService.publishRecordRemoved(oldEntity.getId(), oldEntity);
+  @Override
+  protected Future<List<Pair<String, Instance>>> getInstanceIds(Collection<Instance> instances) {
+    return succeededFuture(instances.stream()
+      .map(instance -> pair(instance.getId(), instance))
+      .collect(Collectors.toList()));
   }
 
-  public Future<Void> publishAllInstancesRemoved() {
-    return domainEventService.publishAllRecordsRemoved();
+  @Override
+  protected Instance convertDomainToEvent(String instanceId, Instance domain) {
+    return domain;
   }
 
-  public Function<Response, Future<Response>> publishInstancesCreatedOrUpdated(
-    BatchOperationContext<Instance> batchOperation) {
-
-    return response -> {
-      if (!isCreateSuccessResponse(response)) {
-        log.warn("Instance create/update failed, skipping event publishing");
-        return succeededFuture(response);
-      }
-
-      log.info("Instances created {}, instances updated {}",
-        batchOperation.getRecordsToBeCreated().size(),
-        batchOperation.getExistingRecords().size());
-
-      return publishInstancesCreated(batchOperation.getRecordsToBeCreated())
-        .compose(notUsed -> publishInstancesUpdated(batchOperation.getExistingRecords()))
-        .map(response);
-    };
-  }
-
-  public Future<Void> publishInstanceReindex(String instanceId, String reindexJobId) {
-    var jobIdHeader = Map.of(REINDEX_JOB_ID_HEADER, reindexJobId);
-
-    return domainEventService.publishMessage(instanceId,
-      reindexEvent(domainEventService.tenantId), jobIdHeader);
-  }
-
-  private Future<Void> publishInstancesUpdated(Collection<Instance> oldInstances) {
-    log.info("[{}] instances were updated, sending events for them", oldInstances.size());
-
-    return instanceRepository.getById(oldInstances, Instance::getId)
-      .map(updatedInstances -> mapOldInstancesToUpdated(updatedInstances, oldInstances))
-      .compose(domainEventService::publishRecordsUpdated);
-  }
-
-  private List<Triple<String, Instance, Instance>> mapOldInstancesToUpdated(
-    Map<String, Instance> updatedInstancesMap, Collection<Instance> oldInstances) {
-
-    return oldInstances.stream()
-      .map(instance -> {
-        final String instanceId = instance.getId();
-        final Instance newInstance = updatedInstancesMap.get(instanceId);
-
-        return new ImmutableTriple<>(instanceId, instance, newInstance);
-      }).collect(Collectors.toList());
+  @Override
+  protected String getId(Instance record) {
+    return record.getId();
   }
 }