+++ src/main/java/org/folio/services/kafka/KafkaProperties.java
@@ -1,23 +1,17 @@
 package org.folio.services.kafka;
 
-import static java.lang.Integer.parseInt;
 import static java.lang.Short.parseShort;
 import static java.lang.System.getenv;
 
-import java.util.Map;
-import org.apache.kafka.common.serialization.StringSerializer;
-
 public final class KafkaProperties {
-  private static int port = parseInt(getenv().getOrDefault("KAFKA_PORT", "9092"));
+  private static String port = getenv().getOrDefault("KAFKA_PORT", "9092");
   private static String host = getenv().getOrDefault("KAFKA_HOST", "localhost");
 
   public static void setPort(int port) {
-    KafkaProducerServiceFactory.clear();
-    KafkaProperties.port = port;
+    KafkaProperties.port = String.valueOf(port);
   }
 
   public static void setHost(String host) {
-    KafkaProducerServiceFactory.clear();
     KafkaProperties.host = host;
   }
 
@@ -25,20 +19,11 @@ public final class KafkaProperties {
     return parseShort(getenv().getOrDefault("REPLICATION_FACTOR", "1"));
   }
 
-  private static String getHost() {
+  public static String getPort() {
+    return port;
+  }
+
+  public static String getHost() {
     return host;
-  }
-
-  public static String getServerAddress() {
-    return getHost() + ":" + port;
-  }
-
-  public static Map<String, String> getProducerProperties() {
-    return Map.of(
-      "bootstrap.servers", getServerAddress(),
-      "enable.idempotence", "true",
-      "key.serializer", StringSerializer.class.getName(),
-      "value.serializer", StringSerializer.class.getName()
-    );
   }
 }