+++ NEWS.md
@@ -1,4 +1,20 @@
-## 20.1.0 In-progress
+## 20.3.0 IN-PROGRESS
+
+* Adds full-text index for classifications field in instance table (MODINVSTOR-716)
+* Adds `effective location` attribute to holdings record and inventory-hierarchy API (MODINVSTOR-669, MODINVSTOR-670)
+* Defaults instance.previouslyHeld to false (MODINVSTOR-454)
+* Provides `holdings-storage 4.6` (MODINVSTOR-669
+* Provides `instance-storage 7.7`
+* Provides `inventory-hierarchy 0.2` (MODINVSTOR-670)
+
+## 20.2.0 2021-04-23
+
+* Determines shelving order for existing items during tenant upgrade (MODINVSTOR-521)
+
+## 20.1.0 2021-03-25
+
+* Introduces `contributorsNames` CQL index for searching `instances` (MODINVSTOR-705)
+* Introduces `REPLICATION_FACTOR` environment variable for configuring Kafka topic replication (MODINVSTOR-694)
 
 ## 20.0.0 2021-03-11